requirements.txt
Django==2.2
django-allauth==0.41.0
django-bootstrap4==1.1.1

django-static-precompiler==1.8.2
libsass==0.19.4

mysqlclient==1.4.6

sendgrid==6.1.0

django-debug-toolbar==2.1

sentry-sdk==0.14.0

# awsebcli==3.15.3
# asgiref==3.2.3
# beautifulsoup4==4.8.2
# certifi==2019.11.28
# chardet==3.0.4
# defusedxml==0.6.0
# idna==2.8
# oauthlib==3.1.0
# python-http-client==3.2.1
# python3-openid==3.1.0
# pytz==2019.3
# requests==2.22.0
# requests-oauthlib==1.3.0
# six==1.13.0
# soupsieve==1.9.5
# sqlparse==0.3.0
# urllib3==1.25.7
